Journal app - "My second own attempt" - Deletion feature implemented
Implemented the feature to delete individual entries. In order to implement that feature, I had to learn how to handle exceptions ("try:" and "exception:") and how to remove indivudal list items. It took me quite a while to implement the feature as I had some issues mapping the user input to the correct entry. This was mainly because the entries shown to the user were listed in the opposite order as they were actually in the stored list, and started counting at 1 rather than 0. But I managed in the end :)