Weather app - My own attempt
Here I try to build the weather app with as little help as possible. Managed to do it mostly on my own, but had to look up how to use the modules properly a few times.
Happy to say that I was able to do the most part on my own though :)