Lolcat app - Instruction led
Learning how to make the lolcat app, which downloads a random set of lolcats from a specified url and displays them to the user.
Completed the instructor led walkthrough of making the app, but am experiencing one issue. Under Windows, the subprocess call to open the folder location in the explorer only opens the explorer at the 'documents' directory, but not at the lolcats. Haven't tested this on Linux or OS X.