Number Guessing app - Revision
As I am currently on a plane without internet access, I decided to revise an older lesson. While making the app I got a bit carried away and introduced error handling in case the user entered a non-integer as his guess, or if the number he entered was outside the range.